Refactor(DK-426): useOutsideClick 훅으로 분리
- 특정 요소 외부 클릭 시 동작을 처리하는 훅 구현
